requirements.txt
# FastAPI 및 웹 서버
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain 핵심 패키지
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-postgres>=0.0.1

# 벡터스토어 및 데이터베이스
pgvector>=0.2.4
psycopg2-binary>=2.9.5
psycopg>=3.1.0
asyncpg>=0.29.0
alembic>=1.13.0
sqlalchemy>=2.0.0

# 추가 유틸리티
python-dotenv>=1.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0

# 선택적 의존성 (OpenAI API 사용 시)
openai>=1.0.0

# CPU 기반 로컬 모델 사용 시 (선택적)
# transformers>=4.30.0
# torch>=2.0.0
# langchain-ollama>=0.1.0
